backend/requirements.txt
gitdb==4.0.12
GitPython==3.1.45
smmap==5.0.2
chromadb>=0.5.3
python-dotenv>=1.0.1
fastapi==0.120.0
openai>=1.0.0
